backend/requirements.txt
torch
torchvision
flask
pillow
scikit-learn
numpy
werkzeug